+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Supported nullable Bicep parameters in Deployment cmdlets
 
 ## Version 6.12.1
 * Used utf8 encoding for reading stdout & stderr when invoking Bicep. [#23246]